requirements.txt
joblib
praat-parselmouth
sounddevice
streamlit
streamlit-webrtc
sklearn
pydub
wave